SEQUOIADBMAINSTREAM-2428 全文检索story检视意见修改：（1）日志中，capped类型不使用INT8；（2）dmsStorageUnit构造函数整改；（3）日志按规范整改
git-svn-id: http://192.168.20.11/sequoiadb/trunk@29265 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/cls/clsReplayer.cpp
@@ -421,7 +421,7 @@ namespace engine
             const CHAR *cs = NULL ;
             INT32 pageSize = 0 ;
             INT32 lobPageSize = 0 ;
-            INT8 type = 0 ;
+            INT32 type = 0 ;
             rc = dpsRecord2CSCrt( (CHAR *)recordHeader,
                                   &cs, pageSize, lobPageSize, type ) ;
             if ( SDB_OK != rc )
@@ -877,7 +877,7 @@ namespace engine
             const CHAR *cs = NULL ;
             INT32 pageSize = 0 ;
             INT32 lobPageSize = 0 ;
-            INT8 type = 0 ;
+            INT32 type = 0 ;
             rc = dpsRecord2CSCrt( (const CHAR *)recordHeader,
                                   &cs, pageSize, lobPageSize, type ) ;
             if ( SDB_OK != rc )
